requirements.txt
pandas
torch
torchvision
torchaudio
matplotlib
numpy
scikit-learn
scipy
ipykernel
jupyter
seaborn
PyYAML